models/aprobaciones_gmm_2026/requirements.txt
pandas
requests
pyyaml
scikit-learn
matplotlib
seaborn
openpyxl
plotly
scipy
